requirement.txt
uvicorn
fastapi
dotenv
huggingface_hub
newsapi-python

